brain/CODE/Brain Tumor Detection/Brain Tumor Detection/requirements.txt
absl-py==1.3.0
asgiref==3.5.2
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
Django==4.1.2
flatbuffers==22.9.24
fonttools==4.37.4
gast==0.4.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
idna==3.4
imutils==0.5.4
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.2
seaborn==0.12.1
six==1.16.0
sqlparse==0.4.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
typing_extensions==4.4.0
tzdata==2022.5
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
